# services/tracker/openapi.json
{
  "components": {
    "schemas": {
...
      "DeletedSandboxEvidence": {
        "additionalProperties": false,
        "properties": {
          "instance_id": {
            "title": "Instance Id",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SandboxSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "const": "deleted",
            "default": "deleted",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "instance_id",
          "snapshot"
        ],
        "title": "DeletedSandboxEvidence",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LiveSandboxEvidence": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "instance_id": {
            "title": "Instance Id",
            "type": "string"
          },
          "last_activity_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Activity At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "resources": {
            "$ref": "#/components/schemas/SandboxResources"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "status": {
            "const": "live",
            "default": "live",
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "provider",
          "instance_id",
          "name",
          "state",
          "region",
          "resources",
          "created_at",
          "updated_at",
          "last_activity_at"
        ],
        "title": "LiveSandboxEvidence",
        "type": "object"
      },
...
      "NotRecordedSandboxEvidence": {
        "additionalProperties": false,
        "properties": {
          "instance_id": {
            "title": "Instance Id",
            "type": "null"
          },
          "message": {
            "maxLength": 200,
            "title": "Message",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "status": {
            "const": "not_recorded",
            "default": "not_recorded",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "message"
        ],
        "title": "NotRecordedSandboxEvidence",
        "type": "object"
      },
...
      "RunLogEvent": {
        "additionalProperties": false,
        "properties": {
          "attempt_id": {
            "title": "Attempt Id",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "ingestion_time_ms": {
            "title": "Ingestion Time Ms",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "timestamp_ms": {
            "title": "Timestamp Ms",
            "type": "integer"
          }
        },
        "required": [
          "event_id",
          "task_id",
          "attempt_id",
          "timestamp_ms",
          "ingestion_time_ms",
          "message"
        ],
        "title": "RunLogEvent",
        "type": "object"
      },
      "RunLogEventsResponse": {
        "additionalProperties": false,
        "properties": {
          "at_tail": {
            "title": "At Tail",
            "type": "boolean"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/RunLogEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "next_cursor": {
            "title": "Next Cursor",
            "type": "string"
          }
        },
        "required": [
          "events",
          "next_cursor",
          "at_tail",
          "is_active"
        ],
        "title": "RunLogEventsResponse",
        "type": "object"
      },
...
      "SandboxResourceSnapshot": {
        "additionalProperties": false,
        "properties": {
          "cpu_cores": {
            "minimum": 0.0,
            "title": "Cpu Cores",
            "type": "number"
          },
          "disk_gib": {
            "minimum": 0.0,
            "title": "Disk Gib",
            "type": "number"
          },
          "gpu_count": {
            "minimum": 0.0,
            "title": "Gpu Count",
            "type": "number"
          },
          "gpu_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Type"
          },
          "memory_gib": {
            "minimum": 0.0,
            "title": "Memory Gib",
            "type": "number"
          }
        },
        "required": [
          "cpu_cores",
          "memory_gib",
          "disk_gib",
          "gpu_count",
          "gpu_type"
        ],
        "title": "SandboxResourceSnapshot",
        "type": "object"
      },
      "SandboxResources": {
        "additionalProperties": false,
        "properties": {
          "cpu_cores": {
            "title": "Cpu Cores",
            "type": "number"
          },
          "disk_gib": {
            "title": "Disk Gib",
            "type": "number"
          },
          "gpu_count": {
            "title": "Gpu Count",
            "type": "number"
          },
          "gpu_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Type"
          },
          "memory_gib": {
            "title": "Memory Gib",
            "type": "number"
          }
        },
        "required": [
          "cpu_cores",
          "memory_gib",
          "disk_gib",
          "gpu_count",
          "gpu_type"
        ],
        "title": "SandboxResources",
        "type": "object"
      },
      "SandboxSnapshot": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "resources": {
            "$ref": "#/components/schemas/SandboxResourceSnapshot"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "name",
          "state",
          "region",
          "resources",
          "recorded_at"
        ],
        "title": "SandboxSnapshot",
        "type": "object"
      },
...
      "UnavailableSandboxEvidence": {
        "additionalProperties": false,
        "properties": {
          "instance_id": {
            "title": "Instance Id",
            "type": "string"
          },
          "message": {
            "maxLength": 200,
            "title": "Message",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SandboxSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "const": "unavailable",
            "default": "unavailable",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "instance_id",
          "message",
          "snapshot"
        ],
        "title": "UnavailableSandboxEvidence",
        "type": "object"
      },
...
      "UnsupportedSandboxEvidence": {
        "additionalProperties": false,
        "properties": {
          "instance_id": {
            "title": "Instance Id",
            "type": "string"
          },
          "message": {
            "maxLength": 200,
            "title": "Message",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SandboxSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "const": "unsupported",
            "default": "unsupported",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "instance_id",
          "message",
          "snapshot"
        ],
        "title": "UnsupportedSandboxEvidence",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "DescopeAccessKey": {
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      },
      "DescopeBearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/benchmarks/{benchmark_id}/logs/events": {
      "get": {
        "description": "Page interleaved task logs for one run, oldest first.",
        "operationId": "get_benchmark_log_events",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Benchmark Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 4096,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunLogEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Get Benchmark Log Events"
      }
    },
...
    "/benchmarks/{benchmark_id}/tasks/{task_id}/attempts/{attempt_id}/sandbox": {
      "get": {
        "description": "Return persisted sandbox identity plus safe live provider metadata.",
        "operationId": "get_task_attempt_sandbox",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Benchmark Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 1,
              "pattern": "^[0-9a-f]+$",
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "deleted": "#/components/schemas/DeletedSandboxEvidence",
                      "live": "#/components/schemas/LiveSandboxEvidence",
                      "not_recorded": "#/components/schemas/NotRecordedSandboxEvidence",
                      "unavailable": "#/components/schemas/UnavailableSandboxEvidence",
                      "unsupported": "#/components/schemas/UnsupportedSandboxEvidence"
                    },
                    "propertyName": "status"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/LiveSandboxEvidence"
                    },
                    {
                      "$ref": "#/components/schemas/DeletedSandboxEvidence"
                    },
                    {
                      "$ref": "#/components/schemas/UnsupportedSandboxEvidence"
                    },
                    {
                      "$ref": "#/components/schemas/UnavailableSandboxEvidence"
                    },
                    {
                      "$ref": "#/components/schemas/NotRecordedSandboxEvidence"
                    }
                  ],
                  "title": "Response Get Task Attempt Sandbox"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Get Task Attempt Sandbox"
      }
    },
...
  }
}